refactor move schema.js to composables
